+++ examples/vue-apollo/src/pages/IndexPage.vue
@@ -1,0 +1,8 @@
+<template>
+  <div className="d-flex align-center flex-column">
+    <v-card width="400">
+      <v-card-title>Index</v-card-title>
+      <v-card-text> Welcome </v-card-text>
+    </v-card>
+  </div>
+</template>