+++ examples/vue-apollo/src/pages/sign-in/EmailPasswordless.vue
@@ -1,0 +1,12 @@
+<template>
+  <email-passwordless />
+  <v-btn
+    class="my-1"
+    block
+    variant="text"
+    color="primary"
+    to="/signin"
+  >
+    &#8592; Other Login Options
+  </v-btn>
+</template>