+++ examples/vue-apollo/vite.config.js
@@ -1,0 +1,20 @@
+import { defineConfig } from 'vite'
+
+import vue from '@vitejs/plugin-vue'
+import vuetify from '@vuetify/vite-plugin'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [
+    vue(),
+    vuetify({
+      autoImport: true
+    })
+  ]
+  // define: { 'process.env': {} },
+  // resolve: {
+  //   alias: {
+  //     '@': path.resolve(__dirname, 'src')
+  //   }
+  // }
+})