+++ packages/core/src/client.ts
@@ -27,7 +27,7 @@ export class AuthClient {
     start = true,
     backendUrl,
     clientUrl,
-
+    devTools,
     ...defaultOptions
   }: NhostClientOptions) {
     this.backendUrl = backendUrl
@@ -43,7 +43,7 @@ export class AuthClient {
     })
 
     if (start) {
-      this.interpreter = interpret(this.machine)
+      this.interpreter = interpret(this.machine, { devTools })
       this.interpreter.start()
     }
 
@@ -56,7 +56,7 @@ export class AuthClient {
         this._channel.addEventListener('message', (token) => {
           const existingToken = this.interpreter?.state.context.refreshToken.value
           if (this.interpreter && token.data !== existingToken) {
-            this.interpreter.send({ type: 'TRY_TOKEN', token: token.data })
+            this.interpreter.send('TRY_TOKEN', { token: token.data })
           }
         })
       } catch (error) {