+++ packages/core/tests/smsOtpSignIn.test.ts
@@ -51,7 +51,7 @@ test(`should fail if network is unavailable`, async () => {
   })
 
   const state: AuthState = await waitFor(authService, (state: AuthState) =>
-    state.matches({ authentication: { signedOut: { failed: 'server' } } })
+    state.matches('authentication.signedOut.failed')
   )
 
   expect(state.context.errors).toMatchInlineSnapshot(`
@@ -75,7 +75,7 @@ test(`should fail if server returns an error`, async () => {
   })
 
   const state: AuthState = await waitFor(authService, (state: AuthState) =>
-    state.matches({ authentication: { signedOut: { failed: 'server' } } })
+    state.matches('authentication.signedOut.failed')
   )
 
   expect(state.context.errors).toMatchInlineSnapshot(`
@@ -98,12 +98,18 @@ test(`should fail if the provided phone number was invalid`, async () => {
   })
 
   const state: AuthState = await waitFor(authService, (state: AuthState) =>
-    state.matches({
-      authentication: { signedOut: { failed: { validation: 'phoneNumber' } } }
-    })
+    state.matches('authentication.signedOut.failed')
   )
 
-  expect(state.context.errors).toMatchObject({ authentication: INVALID_PHONE_NUMBER_ERROR })
+  expect(state.context.errors).toMatchInlineSnapshot(`
+      {
+        "authentication": {
+          "error": "invalid-phone-number\",
+          "message": "Phone number is incorrectly formatted\",
+          "status": 10,
+        },
+      }
+    `)
 })
 
 test(`should fail if the provided OTP was invalid`, async () => {
@@ -116,7 +122,7 @@ test(`should fail if the provided OTP was invalid`, async () => {
   })
 
   const state: AuthState = await waitFor(authService, (state: AuthState) =>
-    state.matches({ authentication: { signedOut: { failed: 'server' } } })
+    state.matches('authentication.signedOut.failed')
   )
 
   expect(state.context.errors).toMatchInlineSnapshot(`