+++ packages/vue/src/useUserIsAnonymous.ts
@@ -1,0 +1,22 @@
+import { useSelector } from '@xstate/vue'
+
+import { useAuthInterpreter } from './useAuthInterpreter'
+
+/**
+ * Use the composable `useUserIsAnonymous` to see if the user is anonymous or not.
+ *
+ * @example
+ * ```tsx
+ * const userIsAnonymous = useUserIsAnonymous();
+ * ```
+ *
+ * @docs https://docs.nhost.io/reference/vue/use-user-is-anonymous
+ */
+export const useUserIsAnonymous = () => {
+  const service = useAuthInterpreter()
+  return useSelector(
+    service.value,
+    (state) => state.context.user?.isAnonymous,
+    (a, b) => a === b
+  )
+}